requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
python-jose[cryptography]
pandas
boto3
pytz